Moving Cart Manager to Entity\Cart
